更新开发环境自动初始化脚本

+++ develop/init.sh
@@ -10,6 +10,7 @@
 # 请在CentOS7服务器上登录root用户运行此脚本
 # 执行过程中需要获取网络资源，请提前配置网络及固定ipv4地址
 # 为避免执行过程中由于网络不通等其他原因导致执行失败，请在执行此脚本前创建虚拟机快照
+# 脚本仍不完善，没有对失败的情况做补偿，经过测试，失败率较低，若执行失败，请手动恢复快照，重新执行
 # 此脚本只需要成功执行一次，虚拟机重启之后只需重新启动docker和容器（默认自启动）
 
 # 检查当前是否root用户
@@ -18,15 +19,22 @@ if [[ `whoami` != "root" ]];then
   exit 1
 fi
 
-# 创建demofile用户
-if [[ `cat /etc/passwd | cut -f1 -d':' | grep -w "demofile" -c` -le 0 ]];then
-  adduser demofile
-  echo "demo" | passwd --stdin demofile
+# 确认提示
+echo "自动初始化脚本即将开始执行，是否确认继续"
+read -p "请输入yes或no: " -r confirm
+if [[ $confirm != "yes" ]];then
+  echo "退出安装"
+  exit 1
 fi
 
-# 关闭防火墙
-systemctl stop firewalld
-systemctl disable firewalld
+# 安装wget
+yum -y install wget
+
+# 修改yum源
+\mv /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.bak
+wget -O /etc/yum.repos.d/CentOS-Base.repo http://mirrors.aliyun.com/repo/Centos-7.repo
+yum clean all
+yum makecache fast
 
 # 安装net-tools
 yum -y install net-tools
@@ -51,7 +59,43 @@ do
   read -p "docker镜像加速地址:" -r docker_mirror
 done
 
-# 准备文件
+echo "开始安装..."
+sleep 3
+
+# 安装docker
+# 安装必要的一些系统工具
+yum install -y yum-utils device-mapper-persistent-data lvm2
+# 添加软件源信息
+yum-config-manager --add-repo http://mirrors.aliyun.com/docker-ce/linux/centos/docker-ce.repo
+# 安装docker-ce
+yum -y install docker-ce
+# 开启docker服务
+systemctl start docker
+
+# 配置docker镜像加速
+mkdir -p /etc/docker
+tee /etc/docker/daemon.json <<-'EOF'
+{
+  "registry-mirrors": ["docker_mirror"]
+}
+EOF
+# 替换为真实镜像地址
+sed -i "s#docker_mirror#$docker_mirror#g" /etc/docker/daemon.json
+
+# 暴露2375端口
+sed -i \
+'s#ExecStart=/usr/bin/dockerd -H fd://#ExecStart=/usr/bin/dockerd -H fd:// -H tcp://0.0.0.0:2375#g' \
+/usr/lib/systemd/system/docker.service
+
+# 重启docker服务
+systemctl daemon-reload
+systemctl restart docker
+
+# docker开机自启动
+systemctl enable docker
+
+sleep 5
+
 # mysql配置文件
 mkdir -p /etc/mysql
 tee /etc/mysql/my.cnf <<-'EOF'
@@ -135,9 +179,26 @@ http {
 EOF
 
 # 容器脚本
+# mysql
+tee mysql.sh <<-'EOF'
+#!/bin/bash
+docker pull mysql:5.7.26
+docker stop mysql && docker rm mysql
+docker run -d --name mysql \
+--restart=always \
+-e MYSQL_ROOT_PASSWORD=root \
+-v /etc/localtime:/etc/localtime \
+-v /etc/timezone:/etc/timezone \
+-v /etc/mysql:/etc/mysql/conf.d \
+-p 3306:3306 \
+mysql:5.7.26
+EOF
+
 # nginx
 tee nginx.sh <<-'EOF'
 #!/bin/bash
+docker pull nginx:1.17-alpine
+docker stop nginx && docker rm nginx
 docker run -d --name nginx \
 --restart=always \
 -v /etc/nginx/nginx.conf:/etc/nginx/nginx.conf \
@@ -152,6 +213,8 @@ EOF
 # redis
 tee redis.sh <<-'EOF'
 #!/bin/bash
+docker pull redis:5-alpine
+docker stop redis && docker rm redis
 docker run -d --name redis \
 --restart=always \
 -v /opt/redis:/data \
@@ -166,6 +229,8 @@ EOF
 # zookeeper
 tee zookeeper.sh <<-'EOF'
 #!/bin/bash
+docker pull wurstmeister/zookeeper
+docker stop zookeeper && docker rm zookeeper
 docker run -d --name zookeeper \
 --restart=always \
 -v /etc/localtime:/etc/localtime \
@@ -178,6 +243,8 @@ EOF
 # kafka
 tee kafka.sh <<-'EOF'
 #!/bin/bash
+docker pull wurstmeister/kafka:2.11-2.0.1
+docker stop kafka && docker rm kafka
 docker run  -d --name kafka \
 --restart=always \
 --add-host docker:local_ip \
@@ -195,73 +262,29 @@ EOF
 # 替换为真实ip
 sed -i "s#--add-host docker:local_ip#--add-host docker:$local_ip#g" kafka.sh
 
-# mysql
-tee mysql.sh <<-'EOF'
-#!/bin/bash
-docker run -d --name mysql \
---restart=always \
--e MYSQL_ROOT_PASSWORD=root \
--v /etc/localtime:/etc/localtime \
--v /etc/timezone:/etc/timezone \
--v /etc/mysql:/etc/mysql/conf.d \
--p 3306:3306 \
-mysql:5.7.26
-EOF
-
-# 安装docker
-# 安装必要的一些系统工具
-yum install -y yum-utils device-mapper-persistent-data lvm2
-# 添加软件源信息
-yum-config-manager --add-repo http://mirrors.aliyun.com/docker-ce/linux/centos/docker-ce.repo
-# 本地yum缓存
-yum makecache fast
-# 安装docker-ce
-yum -y install docker-ce
-# 开启docker服务
-systemctl start docker
-# docker开机自启动
-systemctl enable docker
-
-# 配置docker镜像加速
-mkdir -p /etc/docker
-tee /etc/docker/daemon.json <<-'EOF'
-{
-  "registry-mirrors": ["docker_mirror"]
-}
-EOF
-# 替换为真实镜像地址
-sed -i "s#docker_mirror#$docker_mirror#g" /etc/docker/daemon.json
-
-# 暴露2375端口
-sed -i \
-'s#ExecStart=/usr/bin/dockerd -H fd://#ExecStart=/usr/bin/dockerd -H fd:// -H tcp://0.0.0.0:2375#g' \
-/usr/lib/systemd/system/docker.service
-
-# 重启docker服务
-systemctl daemon-reload
-systemctl restart docker
-
-sleep 5
-
 # 设置时区
 rm -rf /etc/timezone
 echo "Asia/Shanghai" > /etc/timezone
 
-# 下载镜像，创建并启动容器
-docker pull mysql:5.7.26
+# 创建demofile用户
+if [[ `cat /etc/passwd | cut -f1 -d':' | grep -w "demofile" -c` -le 0 ]];then
+  adduser demofile
+  echo "demo" | passwd --stdin demofile
+fi
+
+# 创建并启动容器
 sh mysql.sh
-docker pull nginx:1.17-alpine
 sh nginx.sh
-docker pull redis:5-alpine
 sh redis.sh
-docker pull wurstmeister/zookeeper
 sh zookeeper.sh
-docker pull wurstmeister/kafka:2.11-2.0.1
 sh kafka.sh
-docker pull anapsix/alpine-java:8_server-jre_unlimited
+
+# 关闭防火墙
+systemctl stop firewalld
+systemctl disable firewalld
 
 # 创建并初始化数据库
 docker cp init.sql mysql:/root/
-docker exec -i mysql mysql -uroot -proot -e "create database demo character set utf8mb4;"
-docker exec -i mysql mysql -uroot -proot -e "alter database demo default character set utf8mb4 collate utf8mb4_bin;"
-docker exec -i mysql mysql -uroot -proot demo -e "source /root/init.sql;"
+docker exec -i mysql mysql -uroot -proot -e "source /root/init.sql;"
+
+echo "安装完成"
